[CP-3312] clear-test-files command added
