+++ .gitignore
@@ -65,4 +65,4 @@ apps/mudita-center/static/sql-wasm*
 .nx
 
 matomo-to-gsheet-*.json
-scripts/file-manager-test-files/*
+scripts/manage-test-files/file-manager-test-files/*
